refactor: rename utils/util/ → utils/common/
