+++ commands/agent/delete-policy.js
@@ -1,5 +1,5 @@
 import { deletePolicy } from "../../utils/wallet/keystore.js";
-import { print, printError } from "../../utils/util/output.js";
+import { print, printError } from "../../utils/common/output.js";
 
 export default async function agentDeletePolicy(args, flags) {
   const id = flags.id || flags.name || args[0];
