+++ commands/agent/show-policy.js
@@ -1,6 +1,6 @@
 import { getPolicy } from "../../utils/wallet/keystore.js";
-import { print, printError } from "../../utils/util/output.js";
-import { formatPolicyRules, shortenScriptPaths } from "../../utils/util/format.js";
+import { print, printError } from "../../utils/common/output.js";
+import { formatPolicyRules, shortenScriptPaths } from "../../utils/common/format.js";
 
 export default async function agentShowPolicy(args, flags) {
   const id = flags.id || flags.name || args[0];
