+++ commands/trading/chains.js
@@ -1,6 +1,6 @@
 import { SUPPORTED_CHAINS, getChain } from "../../utils/chain/registry.js";
-import { print } from "../../utils/util/output.js";
-import { formatChains } from "../../utils/util/format.js";
+import { print } from "../../utils/common/output.js";
+import { formatChains } from "../../utils/common/format.js";
 
 export default async function chains(_args, _flags) {
   const chainList = SUPPORTED_CHAINS.map((id) => {
