+++ commands/wallet/sign-message.js
@@ -1,5 +1,5 @@
 import * as ows from "../../utils/wallet/keystore.js";
-import { print, printError } from "../../utils/util/output.js";
+import { print, printError } from "../../utils/common/output.js";
 import { getConfigValue } from "../../utils/config.js";
 import { requireAgentToken } from "../../utils/trading/guards.js";
 import { toCaip2, SUPPORTED_CHAINS } from "../../utils/chain/registry.js";
