+++ commands/wallet/sync.js
@@ -1,5 +1,5 @@
 import * as ows from "../../utils/wallet/keystore.js";
-import { print, printError, isPrettyMode } from "../../utils/util/output.js";
+import { print, printError, isPrettyMode } from "../../utils/common/output.js";
 import { getConfigValue } from "../../utils/config.js";
 
 export default async function walletSync(args, flags) {
