+++ utils/common/analyze.js
